+++ UnityUtils/Packages/com.mydimons.unityutils/Runtime/Script Utils/Audio/MusicClip.cs
@@ -2,26 +2,20 @@ using UnityEngine;
 
 namespace UnityUtils.ScriptUtils.Audio {
   [System.Serializable]
-  public abstract class MusicClip {
+  public abstract class MusicClip : ScriptableObject {
     /// <summary>
     /// The music clip played.
     /// </summary>
     public AudioClip musicClip;
+
     /// <summary>
     /// The name of this music clip.
     /// </summary>
-    public string name;
+    public string musicName = "Music Clip [UnityUtils.ScriptUtils.Audio]";
 
     [Header("Audio Settings")]
-    /// <summary>
-    /// The volume this music clip gets played at 
-    /// </summary>
-    [Range(0, 1)]
-    public float volume = 1;
-    /// <summary>
-    /// The pitch this music clip gets played at
-    /// </summary>
-    public float pitch = 1;
+    public SFXVolumeSettings volumeSettings = new SFXVolumeSettings().SetVolume(1, AudioManager.VolumeType.Music);
+    public SFXPitchSettings pitchSettings = new SFXPitchSettings().SetPitch(0, 1);
 
     /// <returns>If true, the clip can be played when picking music</returns>
     public abstract bool CanBePlayed();